+++ app/MyFireNumber/Views/HealthcareGapPage.xaml.cs
@@ -1,0 +1,12 @@
+using MyFireNumber.ViewModels;
+
+namespace MyFireNumber.Views;
+
+public partial class HealthcareGapPage : CalculatorPageBase
+{
+    public HealthcareGapPage(HealthcareGapViewModel viewModel)
+    {
+        InitializeComponent();
+        InitializeCalculator(viewModel);
+    }
+}
