+++ app/MyFireNumber/Views/SavingsInvestmentPage.xaml.cs
@@ -1,0 +1,12 @@
+using MyFireNumber.ViewModels;
+
+namespace MyFireNumber.Views;
+
+public partial class SavingsInvestmentPage : CalculatorPageBase
+{
+    public SavingsInvestmentPage(SavingsInvestmentViewModel viewModel)
+    {
+        InitializeComponent();
+        InitializeCalculator(viewModel);
+    }
+}
